Progress save: create new class of methods to calculate interception point.
